anpr-yolo/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.1.10           # via tensorflow
appnope==0.1.0            # via ipython
astor==0.6.2              # via tensorflow
bleach==1.5.0             # via tensorboard
cycler==0.10.0            # via matplotlib
decorator==4.2.1          # via ipython, networkx, traitlets
gast==0.2.0               # via tensorflow
grpcio==1.10.0            # via tensorflow
h5py==2.7.1
html5lib==0.9999999       # via bleach, tensorboard
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
jedi==0.11.1              # via ipython
keras==2.1.4
markdown==2.6.11          # via tensorboard
matplotlib==2.1.2         # via scikit-image
networkx==2.1             # via scikit-image
numpy==1.14.1             # via h5py, keras, matplotlib, pywavelets, scipy, tensorboard, tensorflow
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.0.0
prompt-toolkit==1.0.15    # via ipython
protobuf==3.5.1           # via grpcio, tensorboard, tensorflow
ptyprocess==0.5.2         # via pexpect
pygments==2.2.0           # via ipython
pyparsing==2.2.0          # via matplotlib
python-dateutil==2.6.1    # via matplotlib
pytz==2018.3              # via matplotlib
pywavelets==0.5.2         # via scikit-image
pyyaml==3.12              # via keras
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0              # via keras, scikit-image
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via absl-py, bleach, cycler, grpcio, h5py, html5lib, keras, matplotlib, prompt-toolkit, protobuf, python-dateutil, scikit-image, tensorboard, tensorflow, traitlets
tensorboard==1.6.0        # via tensorflow
tensorflow==1.6.0
termcolor==1.1.0          # via tensorflow
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1          # via tensorboard
wheel==0.30.0             # via tensorboard, tensorflow
xmltodict==0.11.0
